abstract patch helper details
